+++ doc/release_notes.rst
@@ -14,7 +14,8 @@ Upcoming Open-TYNDP Release
 
 **Features**
 
-* Add the TYNDP hydrogen demand as an exogenously set demand (process data https://github.com/open-energy-transition/open-tyndp/pull/169)
+* Added the TYNDP hydrogen demand as an exogenously set demand (process data https://github.com/open-energy-transition/open-tyndp/pull/169)
+* Added missing TYNDP supply technologies using PEMMDB capacities, must-runs and availabilities (conventional thermal generation https://github.com/open-energy-transition/open-tyndp/pull/195)
 
 **Bugfixes and Compatibility**